pyproject.toml
[project]
name = "meeting-summarizer"
version = "0.1.0"
description = "Automatically transcribe, diarize, and summarize meetings using Whisper + Claude"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["whisper", "claude", "meeting", "transcription", "summarization", "nlp"]

dependencies = [
    "openai-whisper>=20250625",
    "pyannote.audio>=3.3",
    "torch>=2.0",
    "anthropic>=0.49",
    "requests>=2.32",
    "python-dotenv>=1.0",
]

[project.scripts]
meeting-summarizer = "meeting_summarizer.cli:cli"
download-model     = "meeting_summarizer.download_model:download"

[project.urls]
Repository = "https://github.com/HiIAmTzeKean/meeting-summarizer"
Issues     = "https://github.com/HiIAmTzeKean/meeting-summarizer/issues"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_summarizer"]

# ---------------------------------------------------------------------------
# uv
# ---------------------------------------------------------------------------
[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.11",
]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------
[tool.ruff]
src            = ["src"]
line-length    = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"
